WIP: this doesn't work as of yet.
This code is copied from https://github.com/samvera/hyrax/pull/6241
to solve https://github.com/scientist-softserv/atla-hyku/issues/121. In
that PR, the ability to expire embargoes using a rake task works.

The copied code has not been included in a release yet. At the time of
this writing, September 26, 2023, it is in hyrax main. The latest
release is `hyrax-v4.0.0`.

This app is currently using hyrax-v3.5.0. However, since the changes
being brought in are coming from hyrax main, I'm including the entire
files instead of writing decorators.